This makes master consistent with most of the changes made on bdx_refactoring branch. Notably, it re-adds support for persistent JObjects and JCalibrationManager, although the latter is in need of improvements which will be tackled subsequently. This partly addresses issue #42.

+++ src/lib/JANA/Utils/JCpuInfo.cc
@@ -28,7 +28,7 @@ size_t GetNumCpus() {
 }
 
 
-size_t GetCpuID() {
+uint32_t GetCpuID() {
 	/// Returns the current CPU the calling thread is running on.
 	/// Note that unless the thread affinity has been set, this may 
 	/// change, even before returning from this call. The thread
